requirements_dev.txt
# Generic python dependencies
black==24.4.2
pip==24.3.1
pre-commit==3.6.0
ruff==0.4.4

# Home Assistant testing dependencies
pytest-homeassistant-custom-component==0.13.184
syrupy>=4.6.1

# Component dependencies
hassil<2
home-assistant-intents>=2024.4.3
mutagen>=1.47.0
ha-ffmpeg>=3.2.0
pymicro_vad>=1.0.1
pyspeex_noise>=1.0.2
